+++ app/src/main/java/woowacourse/movie/movie/alarm/SettingPreference.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.movie
+package woowacourse.movie.movie.alarm
 
 import android.content.Context
 import android.content.SharedPreferences
